+++ test/test_helper.rb
@@ -26,13 +26,9 @@ class ActionController::TestCase
 end
 
 class ActionDispatch::IntegrationTest
-  def sign_in_as(user_template, password, email)
-    user = User.create(password: password, password_confirmation: password, email: email,
-                       first_name: user_template.first_name, last_name: user_template.last_name)
-    user.save!
-    user.update_column :deleted, user_template.deleted?
-    user.update_column :system_admin, user_template.system_admin?
-    post_via_redirect '/login', user: { email: email, password: password }
+  def sign_in_as(user, password)
+    user.update password: password, password_confirmation: password
+    post_via_redirect '/login', user: { email: user.email, password: password }
     user
   end
 end